Reorganizing functions
Split functions between three modules: processing, plotting, and fileIO
